restore previous version inside "shape_intrinsic" and add a case handling to the "reshape_intrinsic". The issue was that when a call reshape(x,shape(y)) is made the result information of shape(y) was accessed to find its rank. This howevere is only available if A:  the shape intrinsic executes at least part of the code that had been put 	under an ifdef AND B: one uses either the original result_opnd when processing an intrinsic in call_list_semantics  (see s_call.c:1183)  or keeps using the tmp_result_opnd  and afterwards has to make to copy it back which currently doesn't happen (see the block right after that) because the shape intrinsic generally doesn't have a constant output (even though the rank should always be known and constant).
In general the code uses a tmp_result_opnd and excludes some of the 
intrinsic logic because that logic already tends to lower the representation 
to a point where the reconstruction of the original input is problematic. 
The full shape_intrinsic logic will - aside from determining the rank - 
also determine the dimension as constants if possible and subsequently replace the call to shape OR if the dimensions are variable my lift out the shape call and assign the result to a temporary. Then that temporary 
would replace the argument in reshape BUT the rest of the modified front-end logic would actually ignore the lifted SHAPE so upon unparsing we end up with  an uninitialized temporary being used in RESHAPE. 

Fengmei suggested however that the RESHAPE logic - because further 
optimizations are also prevented- does not actually need the true 
rank value and therefore this fix just adds another condition that sets
rank to zero if it cannot be retrieved from the argument information (that would have been set by the shape_intrinsic logic). 
call to shape