special treatment for pragmas of the shape $openad xxx file_start    <name>
Because of the whirl structure pragmas are not unparsed to a location OUTSIDE of a PU . 
To indicate a file start with the above pragma the working assumption is that the file start has to come just prior to the start of the PU that ends up containing the whirl node for the pragma inside its whirl representation. We simply look ahead in w2f_driver.cxx for the occurence of any such pragma and unparse it first.  Then we unparse the rest of the whirl tree we filter out  in wn2f_pragma these file_start pragma to avoid repeating it.  The "file_start" string is defined in wn2f_pragma.cxx.